feat(FusionReactor): rename package FusionControlComputer to fusion
