+++ src/main/java/techreborn/blocks/generator/BlockFusionControlComputer.java
@@ -38,7 +38,7 @@ import net.minecraft.world.World;
 import reborncore.api.blockentity.IMachineGuiHandler;
 import reborncore.common.blocks.BlockMachineBase;
 import reborncore.common.util.Torus;
-import techreborn.blockentity.machine.multiblock.FusionControlComputer.FusionControlComputerBlockEntity;
+import techreborn.blockentity.machine.multiblock.fusion.FusionControlComputerBlockEntity;
 import techreborn.blockentity.GuiType;
 import techreborn.init.TRContent;
 import techreborn.utils.damageSources.FusionDamageSource;
